styling and saving plots
I created within the results folder a figure folder now, so the plots you are doing can be saved somewhere else in case we want to use for a presentation for example. I removed one of the the libraries that we are not using and added rafalib. I changed the countTable nomenclature based on your counts_raw name. I edited some of the plots to print and save them. But, remember that when we knit the file, we should have an explanation for each plot, right now we have many plots on the same chunk that will not have an explanation. Could you divide one plot for a different chunk with some comment above explaining what is that plot?